requirements_test.txt
coverage
tox
djangorestframework>=3.14
drf-spectacular>=0.27
